CV_Context/frame_context.py
"""
Module for processing computer vision-related questions with frame context.
This module receives questions and the current video frame from the main ADA system,
analyzes the frame based on the question context, and returns a response.
"""

import logging
import os
import time
import cv2

# Set up logging
logging.basicConfig(
    level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
)
logger = logging.getLogger(__name__)

def process_frame_with_context(frame: cv2.Mat, question: str) -> str:
    """
    Process a video frame with the given question context.
    Currently returns a placeholder response.
    
    Args:
        frame: Current video frame to analyze
        question: The question context received from LLM
        
    Returns:
        A string response (currently a placeholder)
    """
    try:
        logger.info(f"Received frame with question context: {question}")
        
        # Create log directory inside the CV_Context folder if it doesn't exist
        log_dir = os.path.join(os.path.dirname(__file__), "logs")
        os.makedirs(log_dir, exist_ok=True)
        
        # Log the received question and frame details
        timestamp = time.strftime("%Y%m%d-%H%M%S")
        log_file = os.path.join(log_dir, f"cv_frame_analysis.log")
        
        frame_height, frame_width = frame.shape[:2]
        
        # Simulate some processing time to avoid immediate response
        # This helps prevent the continuous loop of questions
        time.sleep(1)
        
        # Placeholder response based on question type
        if "age" in question.lower():
            response = "Based on facial features, the person appears to be in their late 20s to early 30s."
        elif "cloth" in question.lower() or "wear" in question.lower():
            response = "The person appears to be wearing casual attire. Specific details would require more advanced cloth detection."
        elif "count" in question.lower() or "people" in question.lower() or "person" in question.lower():
            response = "There is 1 person detected in the current frame."
        else:
            response = "Analysis complete. I've processed the frame according to your request."
        
        # Log the response
        with open(log_file, "a") as f:
            f.write(f"[{timestamp}] Received question: {question}\n")
            f.write(f"[{timestamp}] Frame dimensions: {frame_width}x{frame_height}\n")
            f.write(f"[{timestamp}] Response: {response}\n\n")
        
        logger.info(f"Frame analysis complete. Response: {response}")
        return response
        
    except Exception as e:
        logger.error(f"Error processing frame with context: {e}")
        return "Error processing the frame analysis request"